[chore/#142] enum 이동, VLine viewModel 이동
