+++ ILSANG/Sources/Views/Quest/QuestItemView.swift
@@ -89,13 +89,3 @@ struct QuestItemView: View {
         QuestItemView(quest: QuestViewModelItem.mockData, status: .uncompleted)
     }
 }
-
-
-struct VLine: Shape {
-    func path(in rect: CGRect) -> Path {
-        var path = Path()
-        path.move(to: CGPoint(x: 0, y: 0))
-        path.addLine(to: CGPoint(x: 0, y: rect.height))
-        return path
-    }
-}
